satyanarayan/models.py
from django.db import models
from ckeditor.fields import RichTextField

class Post(models.Model):
    id = models.IntegerField(unique=True, auto_created=True, primary_key=True)
    title = models.CharField(max_length=200)
    thumbnail = models.ImageField(upload_to='thumbnails')
    content = RichTextField()
    def __str__(self):
        return self.title

class Facilitie(models.Model):
    title = models.CharField(max_length=20)
    img = models.ImageField(upload_to='thumbnails')
    def __str__(self):
        return self.title